+++ CHANGELOG.rst
@@ -36,7 +36,10 @@ Breaking changes
  :py:meth:`~climpred.classes.PredictionEnsemble.verify` and
  :py:meth:`~climpred.classes.PredictionEnsemble.bootstrap` to ``initialized``,
   ``uninitialized`` and ``persistence``. Also ``reference`` should be choosen from
-  [``uninitialized``, ``persistence``]. (:issue:`460`, :pr:`478`) `Aaron Spring`_
+  [``uninitialized``, ``persistence``]. (:issue:`460`, :pr:`478`, :issue:`476`,
+  :pr:`480`) `Aaron Spring`_
+- :py:meth:`~climpred.classes.PredictionEnsemble.verify` returns no ``skill`` dimension
+  if ``reference=None``.  (:pr:`480`) `Aaron Spring`_
 
 New Features
 ------------